Refactor: Organize file structure and remove unused code
- Created a `src` directory to house all JavaScript source files, cleaning up the root directory.
- Moved all `.js` files into the new `src` directory.
- Updated all `require` paths in the application and test files to reflect the new structure.
- Updated the `main` and `start` script entries in `package.json`.

- Created a new `src/api-client.js` module to encapsulate all `axios`-based interactions with the Hello Club API.
- Refactored `src/functions.js` to be leaner by removing API logic and delegating calls to the new `api-client`.

- Deleted the unused `csv-generator.js` file.

- Reran `npm install` to resolve module discovery issues within the Jest test environment.
- Fixed and updated all tests to align with the refactored code, ensuring all tests pass.
